refactor: update to ChatQuestionnaire
