+++ src/components/Domain/Jobs/ChatQuestionnaire/ChatQuestionnaire.component.tsx
@@ -1,0 +1,23 @@
+import React, { forwardRef } from 'react'
+import { classNames } from '@shared/utils/common'
+import { Chat, IChat } from '@components/UI/molecules/Chat'
+import { IChatQuestionnaire, IQuestionnaireMessage, IQuestionWithAnswer } from './ChatQuestionnaire.interface'
+import styles from './ChatQuestionnaire.module.scss'
+
+const cx = classNames.bind(styles)
+
+export const ChatQuestionnaire = forwardRef<IChat.Methods, IChatQuestionnaire>(({ children, className }, chat) => {
+
+  return (
+    <Chat className={cx('questionnaire', className)} ref={chat}>
+      {({ messages }) => (
+        <>
+          <div className={cx('questionnaire__mask', { 'questionnaire__mask--hidden': messages.every((msg) => (msg.content as IQuestionWithAnswer).mode === 'readonly') })}></div>
+          {children?.(messages as IQuestionnaireMessage[])}
+        </>
+    )}
+    </Chat>
+  )
+})
+
+ChatQuestionnaire.displayName = 'Questionnaire'
