+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireHeader/ChatQuestionnaireHeader.interface.ts
@@ -1,0 +1,8 @@
+
+export interface IChatQuestionnaireHeader extends React.HTMLAttributes<HTMLElement> {
+    classNames?: {
+        iconBack?: string
+        iconLogo?: string
+        title?: string
+    }   
+}
