+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireHeader/ChatQuestionnaireHeader.stories.tsx
@@ -1,0 +1,16 @@
+import { StoryObj, Meta } from '@storybook/react'
+import { ChatQuestionnaireHeader } from './ChatQuestionnaireHeader.component'
+
+const meta: Meta<typeof ChatQuestionnaireHeader> = {
+  title: 'Domain/Jobs/ChatQuestionnaire/ChatQuestionnaireHeader',
+  component: ChatQuestionnaireHeader
+}
+export default meta
+
+type Story = StoryObj<typeof ChatQuestionnaireHeader>
+
+export const Default: Story = {
+  args: {
+    children: 'Questionnaire Header',
+  }
+}
