+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireHeader/index.ts
@@ -1,0 +1,2 @@
+export { ChatQuestionnaireHeader } from './ChatQuestionnaireHeader.component'
+export type { IChatQuestionnaireHeader } from './ChatQuestionnaireHeader.interface'
