+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireMessageCandidate/ChatQuestionnaireMessageCandidate.module.scss
