+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireMessageCandidate/index.ts
@@ -1,0 +1,1 @@
+export { ChatQuestionnaireMessageCandidate } from './ChatQuestionnaireMessageCandidate.component'
