+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireMessageLazy/ChatQuestionnaireMessageLazy.interface.tsx
@@ -1,0 +1,5 @@
+export interface IChatQuestionnaireMessageLazy {
+  children?: React.ReactNode
+  waitFor?: number
+  onLoad?: VoidFunction
+}
