+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireMessageLazy/index.ts
@@ -1,0 +1,2 @@
+export type { IChatQuestionnaireMessageLazy } from './ChatQuestionnaireMessageLazy.interface'
+export { ChatQuestionnaireMessageLazy } from './ChatQuestionnaireMessageLazy.component'
