+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireMultipleChoice/ChatQuestionnaireMultipleChoice.interface.ts
@@ -1,0 +1,8 @@
+import { IQuestionnaireAnswer } from "../../ChatQuestionnaire.interface";
+
+export interface IChatQuestionnaireMultipleChoice extends IQuestionnaireAnswer {
+    renderSubmitButton: (options: {
+        disabled: boolean;
+        className: string;
+    }) => React.ReactNode;
+}
