+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireMultipleChoice/index.ts
@@ -1,0 +1,2 @@
+export { ChatQuestionnaireMultipleChoice } from './ChatQuestionnaireMultipleChoice.component'
+export type { IChatQuestionnaireMultipleChoice } from './ChatQuestionnaireMultipleChoice.interface'
