+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireQuestion/ChatQuestionnaireQuestion.interface.ts
@@ -1,0 +1,7 @@
+import { IQuestionWithAnswer } from "../../ChatQuestionnaire.interface"
+
+export interface IChatQuestionnaireQuestion {
+  className?: string
+  children?: React.ReactNode
+  questionWithAnswer: IQuestionWithAnswer
+}
