+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireQuestion/index.ts
@@ -1,0 +1,1 @@
+export { ChatQuestionnaireQuestion } from './ChatQuestionnaireQuestion.component'
