+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireUniqueChoice/ChatQuestionnaireUniqueChoice.interface.ts
@@ -1,0 +1,9 @@
+import { IQuestionnaireAnswer } from "../../ChatQuestionnaire.interface"
+
+export interface IChatQuestionnaireUniqueChoice extends IQuestionnaireAnswer {
+  className?: string
+  classNames?: {
+    answers?: string
+    option: string
+  }
+}
