+++ src/components/Domain/Jobs/ChatQuestionnaire/children/ChatQuestionnaireUniqueChoice/index.ts
@@ -1,0 +1,2 @@
+export { ChatQuestionnaireUniqueChoice } from './ChatQuestionnaireUniqueChoice.component'
+export type { IChatQuestionnaireUniqueChoice } from './ChatQuestionnaireUniqueChoice.interface'
