+++ src/components/Domain/Jobs/ChatQuestionnaire/index.ts
@@ -1,0 +1,2 @@
+export { ChatQuestionnaire } from './ChatQuestionnaire.component';
+export type { IChatQuestionnaire, IQuestionWithAnswer, TSendQuestion, TSendCloseQuestion } from './ChatQuestionnaire.interface';
