--- src/components/Domain/Jobs/Questionnaire/children/QuestionnaireHeader/QuestionnaireHeader.interface.ts
@@ -1,8 +1,0 @@
-
-export interface IQuestionnaireHeader extends React.HTMLAttributes<HTMLElement> {
-    classNames?: {
-        iconBack?: string
-        iconLogo?: string
-        title?: string
-    }   
-}
