--- src/components/Domain/Jobs/Questionnaire/children/QuestionnaireHeader/index.ts
@@ -1,2 +1,0 @@
-export { QuestionnaireHeader } from './QuestionnaireHeader.component'
-export type { IQuestionnaireHeader } from './QuestionnaireHeader.interface'
