--- src/components/Domain/Jobs/Questionnaire/children/QuestionnaireMessageCandidate/index.ts
@@ -1,1 +1,0 @@
-export { QuestionnaireMessageCandidate } from './QuestionnaireMessageCandidate.component'
