--- src/components/Domain/Jobs/Questionnaire/children/QuestionnaireMessageLazy/QuestionnaireMessageLazy.interface.tsx
@@ -1,5 +1,0 @@
-export interface IQuestionnaireMessageLazy {
-  children?: React.ReactNode
-  waitFor?: number
-  onLoad?: VoidFunction
-}
