--- src/components/Domain/Jobs/Questionnaire/children/QuestionnaireMessageLazy/index.ts
@@ -1,2 +1,0 @@
-export type { IQuestionnaireMessageLazy } from './QuestionnaireMessageLazy.interface'
-export { QuestionnaireMessageLazy } from './QuestionnaireMessageLazy.component'
