--- src/components/Domain/Jobs/Questionnaire/children/QuestionnaireMultipleChoice/QuestionnaireMultipleChoice.interface.ts
@@ -1,8 +1,0 @@
-import { IQuestionnaireAnswer } from "../../Questionnaire.interface";
-
-export interface IQuestionnaireMultipleChoice extends IQuestionnaireAnswer {
-    renderSubmitButton: (options: {
-        disabled: boolean;
-        className: string;
-    }) => React.ReactNode;
-}
