--- src/components/Domain/Jobs/Questionnaire/children/QuestionnaireMultipleChoice/index.ts
@@ -1,2 +1,0 @@
-export { QuestionnaireMultipleChoice } from './QuestionnaireMultipleChoice.component'
-export type { IQuestionnaireMultipleChoice } from './QuestionnaireMultipleChoice.interface'
