--- src/components/Domain/Jobs/Questionnaire/children/QuestionnaireQuestion/index.ts
@@ -1,1 +1,0 @@
-export { QuestionnaireQuestion } from './QuestionnaireQuestion.component'
