--- src/components/Domain/Jobs/Questionnaire/children/QuestionnaireUniqueChoice/QuestionnaireUniqueChoice.interface.ts
@@ -1,9 +1,0 @@
-import { IQuestionnaireAnswer } from "../../Questionnaire.interface"
-
-export interface IQuestionnaireUniqueChoice extends IQuestionnaireAnswer {
-  className?: string
-  classNames?: {
-    answers?: string
-    option: string
-  }
-}
