--- src/components/Domain/Jobs/Questionnaire/children/QuestionnaireUniqueChoice/index.ts
@@ -1,1 +1,0 @@
-export { QuestionnaireUniqueChoice } from './QuestionnaireUniqueChoice.component'
