--- src/components/Domain/Jobs/Questionnaire/index.ts
@@ -1,2 +1,0 @@
-export { Questionnaire } from './Questionnaire.component';
-export type { IQuestionnaire, IQuestionWithAnswer, TSendQuestion, TSendCloseQuestion } from './Questionnaire.interface';
